backend/requirements-base.txt
# Core
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==22.0.0

# Data
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2

# DB
sqlalchemy==2.0.30
python-multipart==0.0.6

# Auth
python-jose[cryptography]==3.3.0

# Utils
requests==2.32.2
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Testing
pytest==7.4.2
httpx==0.25.0

psycopg2-binary==2.9.9

PyJWT==2.8.0

passlib==1.7.4
bcrypt==4.0.1  # Explicit version